refactor: placed news url provider in single package
