+++ monew/src/main/java/com/codeit/team2/monew/module/domain/article/external/url_provider/HankyungNewsUrlProvider.java
@@ -1,4 +1,4 @@
-package com.codeit.team2.monew.module.domain.article.external;
+package com.codeit.team2.monew.module.domain.article.external.url_provider;
 
 
 import java.util.Set;
