+++ monew/src/test/java/com/codeit/team2/monew/module/domain/article/external/HankyungApiNewsClientTest.java
@@ -1,6 +1,7 @@
 package com.codeit.team2.monew.module.domain.article.external;
 
 import com.codeit.team2.monew.module.domain.article.entity.Article;
+import com.codeit.team2.monew.module.domain.article.external.url_provider.NewsUrlProvider;
 import com.codeit.team2.monew.module.domain.article.mapper.ArticleMapper;
 import com.codeit.team2.monew.module.domain.article.mapper.ArticleMapperImpl;
 import java.io.IOException;
